// app/models/app_service_list.ts
export default class App_service_list {
  ID: string;
  APP_NAME: string;
  ENTITY:string;
  URL: string;
  DESCRIPTION:string; 
  SERVICE_NAME: string;
  IP:string;
  STATUS: string;
  STATUS_CODE: Number;
  FILES_UPLOAD:string;
  STATUS_WEB: string;
  START_TIME: Number;
  END_TIME: Number;
  RESPONSE: Number;
  COLSPAN_APP_NAME:string;
  COLSPAN_IP:string;
}